+++ restaurant-platform-api/src/app.module.ts
@@ -17,6 +17,7 @@ import { FeedbackModule } from './modules/feedback/feedback.module';
 import { AnalyticsModule } from './modules/analytics/analytics.module';
 import { NotificationsModule } from './modules/notifications/notifications.module';
 import { AdminModule } from './modules/admin/admin.module';
+import { SubscriptionsModule } from './modules/subscriptions/subscriptions.module';
 import { WebsocketsModule } from './websockets/websockets.module';
 import { HealthController } from './health.controller';
 import { GlobalExceptionFilter } from './common/filters/http-exception.filter';
@@ -56,6 +57,7 @@ import { JwtAuthGuard } from './common/guards';
     AnalyticsModule,
     NotificationsModule,
     AdminModule,
+    SubscriptionsModule,
     WebsocketsModule,
   ],
   controllers: [HealthController],